refactor: extract DSPy llm so that it can be reused by sanitizer later
(cherry picked from commit df24d814758f66a55df885cab0845dc062b51105)
